Fixed entity editing/undo I'm honestly not 100% sure how the previous code was functioning. It kinda looks like it wasn't implemented properly?!

+++ CaveStoryEditor/Utilities/IUndo.cs
@@ -29,9 +29,9 @@ namespace CaveStoryEditor
     {
         public class EntityPropertyChanged
         {
-            public int OldValue { get; }
-            public int NewValue { get; set; }
-            public EntityPropertyChanged(int old, int @new)
+            public short OldValue { get; }
+            public short NewValue { get; set; }
+            public EntityPropertyChanged(short old, short @new)
             {
                 OldValue = old;
                 NewValue = @new;
